Banner Section
Updated the method that gets the next unused CRN to place on a new Banner section to call a stored procedure to validate the CRN is not used in Banner.  There are cases where Banner sections that are deleted on the UniTime side of the interface are marked as cancelled but not deleted on the Banner side of the interface.  If one of those CRNs is reused by UniTime the section will fail to be loaded into Banner.  This additional validation with Banner ensures a CRN is chosen that is not currently in use for the term within both UniTime and Banner.

To use this functionality a stored procedure to be called within the Banner database must be created.  The call to the stored procedure is defined by the property: 'banner.crnValidator.storedProcedure.call'.  The required format of the parameters for the Banner Stored procedure must be:

         in_term => ?, in_crn => ?, out_used => ?

The names of the parameters may change but the first parameter must be a String input, the second parameter must be an Integer input, and the third parameter must be a String output ('Y' if used in Banner or 'N' if not used in Banner)

An example of how to define the Banner stored procedure call in the properties file is as follows:

banner.crnValidator.storedProcedure.call=begin sz_unitime.p_validate_crn(in_term => ?, in_crn => ?, out_used => ?); end;

The following values must also be defined in the properties:

banner.host=<the name of the host the banner database resides on>
banner.database=<the name of the banner database>
banner.port=<the port used to connect to the database host>
banner.user=<a user with the permission to run the stored procedure>
banner.password=<the user password>

The above values should already be defined in the properties file as they are the same values used by the queue processor to send and retrieve data from Banner.

If the value of "banner.crnValidator.storedProcedure.call" is not set, the validation will always return true and no check will be done to see if the CRN is in use for the term in Banner.