Updates to Build File
Updated build script to attempt figure out the location of core UniTime
code when it is not specified in the build.properties.  This should
allow the build script to work whether the code is checked out via git
or svn.

Also, updated the Banner Add On build to allow the option to build only
a minimal set of gwt permutations.

(cherry picked from commit 93e326dd12bacd1c490582fa27df6e73fd85765a)