MeetingElement - changed handling of alternate date patterns
Whenever a date pattern is marked as alternate rather than sending a
meeting element for each week the class meets, a single meeting element
is now sent with the start date as the first date in the date pattern
and the end date as the last date in the date pattern.  Also, the
ALT_WEEKS attribute has been added which is set to true for alternate
week date patterns and false for all others.