Updated to support additional capabilities of teaching responsibilities.
Updated to support additional capabilities of teaching responsibilities.

- Class Instructors with a teaching responsibility marked as “No
Report” are never sent to Colleague.
- Class Instructors with a null teaching responsibility and 0 percent
responsibility are sent to Colleague with 0 percent responsibility.
- Class Instructors with a null teaching responsibility and a percent
responsibility not equal to 0 have their percent responsibility
normalized so the total responsibility of all instructors is 100% and
the precent responsibility is sent to Colleague.
- Class Instructors with a teaching responsibility that is not marked
as “Auxiliary” and a percent responsibility not equal to 0 have their
percent responsibility normalized so the total responsibility of all
instructors is 100% and the precent responsibility is sent to Colleague.
- Class Instructors with a teaching responsibility that is marked as
“Auxiliary” are sent to Colleague with a percent responsibility of 0 no
matter what percent responsibility is stored on the instructor.
- Course coordinators with a teaching responsibility not marked as “No
Report” or with a null teaching responsibility are sent to Colleague on
every section with a percent responsibility of 0 regardless of the
state of the auxiliary flag on the teaching responsibility .
- Course coordinators with a teaching responsibility marked ad “No
Report” are never sent to Colleague
- Improved normalization of percent teaching responsibility to fix case
where total might not add up to 100.