Script: Subject Areas: Restore Instructors After Department Change
Added a script named 'Subject Areas: Restore Instructors After
Department Change’ that will pull instructors into the new department
for a subject area whenever it changes departments.  It only pulls
instructors that were assigned to classes in the session the classes
were rolled forward from.  It will pull the instructors individual
preferences.  It will also assign the instructors to the same classes
they were assigned to in the last like term.  If a session is marked a
send data to Banner, it will generate an update message to Banner for
each instructional offering.

As a safety net, it can be with the “Do It” flag unchecked and it will
just list the instructors that will be moved and the classes they will
be assigned to. Once the output has been reviewed the output and the
changes are acceptable then check the “Do It” flag and it will actually
make the changes listed.

It does not does not touch instructors or instructor class assignments
that have already been made in the term.

The script should be run in the session where the subject area’s
department has been changed.