Banner Section - Fixed section delete bug
When classes of the same instructional type were nested under each
other, a delete of one of the nested classes or the associated
instructional type would cause the Banner Section to be deleted and
recreated.  This would give the section a new CRN.  This was a problem
when students are enrolled in the original CRN in Banner.  This was
resolved by ensuring the Banner section is only deleted when the last
associated class is deleted.

Some code cleanup was also done.