Colleague Sections - Allow for multiple instructional methods
Updated the Colleague Sections to contain all classes in a nested
structure regardless of instructional type.
Updated the Colleague Message to send the instructional method on the
meetings rather than the section.
Fixed a bug with the deletion of Colleague sections when a class that
is parts of a nested structure is deleted.  This was causing the
original Colleague section to be deleted and a new one created.  This
has been resolved by only marking a Colleague section deleted if all
classes associated with the Colleague section are deleted.