Banner Section: Correct problem with nesting of same instructional types
When a course configuration was created with two or more scheduling
subparts of the same instructional type that were not nested and the
user later edited the course configuration to nest the same
instructional type scheduling subparts, the Banner portion of the
update would fail.  This commit corrects that problem.