Banner Instructional Method Cohort Restriction
Added support to record cohort restrictions on instructional methods and send them to all sections that belong to instructional offering configurations assigned to the instructional method.

This page relies on Instructional Methods being defined using the Administration -> Other -> Instructional Methods page and student groups being defined and set to a type of “COHORT”.  The student group type of “COHORT” can be defined using the Administration -> Other -> Student Group Types page.

There are two properties associated with this feature.

banner.menu.display_inst_method_restrictions - If this property exists and is set to “true” the Administration -> Banner -> Instr Method Restrictions page will become available and the “Restrictions” column will on the Courses -> Banner -> Banner Offerings page.

banner.menu.display_inst_method_restrictions_show_removed - If this property exists and is set to “true” restrictions that have been marked as removed will also be listed in the “Restrictions” column on the Courses -> Banner -> Banner Offerings page.

Once an Instructional Methods is assigned to an Instructional Offering Configuration in an academic session it is no longer possible to delete the Cohort Restrictions.  From that point forward, the restrictions can only be marked as removed.

The tuple of Academic Session, Instructional Method, and Cohort must always be unique.

If a Cohort Restrictions is added or modified for an Instructional Method that is already in use, all sections associated with that Instructional Method must be resent to Banner for the change to take effect.