Meeting Element
compareTo -- Fixed the compareTo method to distinguish between arrange hours meetings that have different rooms.

datePatternDates -- corrected problem with generating date blocks for date patterns that begin long before the start of the term and have 1 or more days where the date pattern does not meet.

datePatternDates -- corrected problem with generating date blocks for date patterns that begin long after the end of the term and have 1 or more days where the date pattern does not meet.

datePatternDates - corrected a problem generating date blocks when a date pattern has a single day during the session where it does not meet.